back out of flattening
