+++ src/test/resources/submit_s1.sh
@@ -2,4 +2,4 @@
 # To use: start RabbitMQ, MongoDB, the FrontendApplication, the Dispatcher, then run this script
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 
-curl localhost:8080/api/domains/testDomain/submissions -H "Content-Type: application/json" -X POST -d "@$DIR/s1.json" -v
+curl localhost:8080/api/submissions -H "Content-Type: application/json" -X POST -d "@$DIR/s1.json" -v
