+++ subs-data-model/src/main/java/uk/ac/ebi/subs/data/submittable/Submittable.java
@@ -1,5 +1,6 @@
 package uk.ac.ebi.subs.data.submittable;
 
+import uk.ac.ebi.subs.data.Submission;
 import uk.ac.ebi.subs.data.component.Attribute;
 import uk.ac.ebi.subs.data.component.Domain;
 import uk.ac.ebi.subs.data.component.Archive;
@@ -7,6 +8,7 @@ import uk.ac.ebi.subs.data.component.Archive;
 import java.util.List;
 
 public interface Submittable {
+
     String getAccession();
     void setAccession(String accession);
