+++ subs-frontend/src/main/java/uk/ac/ebi/subs/frontend/handlers/StudyEventHandler.java
@@ -3,8 +3,9 @@ package uk.ac.ebi.subs.frontend.handlers;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.rest.core.annotation.HandleBeforeCreate;
-import org.springframework.data.rest.core.annotation.RepositoryEventHandler;
+import org.springframework.data.rest.core.annotation.*;
+import org.springframework.stereotype.Component;
+import org.springframework.stereotype.Service;
 import uk.ac.ebi.subs.data.submittable.Study;
 import uk.ac.ebi.subs.repository.submittable.StudyRepository;
 
@@ -15,8 +16,9 @@ import java.util.UUID;
  * Repo event handler for studies nested in a submission
  *
  */
+@Component
 @RepositoryEventHandler(Study.class)
-public class SubmissionStudyEventHandler {
+public class StudyEventHandler {
 
     private final Logger logger = LoggerFactory.getLogger(this.getClass());
 
@@ -29,5 +31,20 @@ public class SubmissionStudyEventHandler {
         study.setId(UUID.randomUUID().toString());
     }
 
+    @HandleBeforeSave
+    public void handleBeforeSave(Study study){
+        logger.warn("save");
+        study.setId(UUID.randomUUID().toString());
+    }
+
+    @HandleAfterCreate
+    public void handleAfterCreate(Study study){
+        logger.warn("aftercreate");
+    }
+
+    @HandleAfterSave
+    public void handleAfterSave(Study study){
+        logger.warn("aftersave");
+    }
 
 }
