+++ subs-frontend/src/main/java/uk/ac/ebi/subs/frontend/handlers/SubmissionEventHandler.java
@@ -4,6 +4,9 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.rest.core.annotation.*;
+import org.springframework.stereotype.Component;
+import org.springframework.stereotype.Service;
+import org.springframework.web.bind.annotation.PathVariable;
 import uk.ac.ebi.subs.data.Submission;
 import uk.ac.ebi.subs.frontend.exceptions.ResourceLockedException;
 import uk.ac.ebi.subs.frontend.services.SubmissionProcessingService;
@@ -19,6 +22,7 @@ import java.util.UUID;
  *  * locks down changes to non-draft submissions, based on the OperationControlService
  *  * send submissions off to rabbit after storing a submission with the 'Submitted' status
  */
+@Component
 @RepositoryEventHandler(Submission.class)
 public class SubmissionEventHandler {
