+++ subs-frontend/src/test/java/uk/ac/ebi/subs/frontend/SubmissionControllerIT.java
@@ -4,38 +4,49 @@ import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.amqp.rabbit.annotation.RabbitListener;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.context.embedded.LocalServerPort;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.web.client.TestRestTemplate;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.repository.CrudRepository;
+import org.springframework.hateoas.Resource;
+import org.springframework.hateoas.client.Traverson;
+import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.web.client.RestTemplate;
 import uk.ac.ebi.subs.FrontendApplication;
+
 import uk.ac.ebi.subs.data.Submission;
+import uk.ac.ebi.subs.processing.ProcessingStatus;
 import uk.ac.ebi.subs.processing.SubmissionEnvelope;
 import uk.ac.ebi.subs.messaging.Queues;
 import uk.ac.ebi.subs.repository.SubmissionRepository;
 import uk.ac.ebi.subs.repository.submittable.*;
 
+import java.net.URI;
 import java.net.URISyntaxException;
-import java.net.URL;
 import java.util.*;
 
 import static org.hamcrest.Matchers.equalTo;
+import static org.hamcrest.Matchers.is;
+import static org.hamcrest.Matchers.notNullValue;
 import static org.junit.Assert.assertThat;
 
 @RunWith(SpringRunner.class)
 @SpringBootTest(classes = FrontendApplication.class, webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 public class SubmissionControllerIT {
 
+    private final Logger logger = LoggerFactory.getLogger(this.getClass());
+
     @LocalServerPort
     private int port;
 
-    private URL submit;
+    private URI submissionsUri;
 
     private List<Submission> submissionsReceived;
 
@@ -92,13 +103,12 @@ public class SubmissionControllerIT {
 
     @Before
     public void setUp() throws Exception {
-        this.submit = new URL("http://localhost:" + this.port + "/api/submit");
-
-
+        this.submissionsUri = URI.create("http://localhost:" + this.port + "/api/submissions");
 
         deleteAllRepos();
 
         sub = new Submission();
+        sub.setStatus(ProcessingStatus.Draft.name());
         sub.getDomain().setName("integrationTestExampleDomain."+UUID.randomUUID().toString());
         sub.getSubmitter().setEmail("test@example.ac.uk");
 
@@ -116,13 +126,22 @@ public class SubmissionControllerIT {
 
     @Test
     public void doSubmit() throws URISyntaxException, InterruptedException {
-        Submission receivedSub = restTemplate.postForObject(submit.toString(), sub, sub.getClass());
+        ResponseEntity<Void> response = restTemplate.postForEntity(submissionsUri.toString(), sub, Void.class);
 
-        Page<Submission> subs = submissionRepository.findAll(new PageRequest(0,100));
-        assertThat(subs.getTotalElements(), equalTo(1L));
+        assertThat(response.getStatusCodeValue(), is(equalTo(201)));
+        assertThat(response.getHeaders().getLocation(),notNullValue());
+
+
+        URI location = response.getHeaders().getLocation();
+
+        ResponseEntity<Resource> submissionResource = restTemplate.postForEntity(
+                submissionsUri.toString(), sub, Resource.class);
+
+
+
 
         Thread.sleep(1000);
 
-        assertThat(submissionsReceived.size(),equalTo(1));
+        assertThat(submissionsReceived.size(),equalTo(0));
     }
 }
