+++ subs-processing-model/src/main/java/uk/ac/ebi/subs/processing/SubmissionEnvelope.java
@@ -2,7 +2,7 @@ package uk.ac.ebi.subs.processing;
 
 
 import org.springframework.data.mongodb.core.mapping.DBRef;
-import uk.ac.ebi.subs.data.FullSubmission;
+
 import uk.ac.ebi.subs.data.Submission;
 import uk.ac.ebi.subs.data.component.SampleRef;
 import uk.ac.ebi.subs.data.submittable.Sample;
@@ -14,7 +14,7 @@ public class SubmissionEnvelope {
 
     String id;
 
-    FullSubmission submission;
+    Submission submission;
 
     Set<SampleRef> supportingSamplesRequired = new HashSet<>();
     List<Sample> supportingSamples = new ArrayList<>();
@@ -23,16 +23,16 @@ public class SubmissionEnvelope {
 
     public SubmissionEnvelope() {};
 
-    public  SubmissionEnvelope(FullSubmission submission){
+    public  SubmissionEnvelope(Submission submission){
         this.submission = submission;
     }
 
 
-    public FullSubmission getSubmission() {
+    public Submission getSubmission() {
         return submission;
     }
 
-    public void setSubmission(FullSubmission submission) {
+    public void setSubmission(Submission submission) {
         this.submission = submission;
     }
