+++ subs-repository/src/main/java/uk/ac/ebi/subs/repository/SubmissionRepository.java
@@ -11,16 +11,12 @@ import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 import org.springframework.data.rest.core.annotation.RestResource;
 import uk.ac.ebi.subs.data.Submission;
 
-@RepositoryRestResource(
-        collectionResourceRel = "submissions",
-        path = "submissions",
-        collectionResourceDescription = @Description("Metadata submissions to the EBI archives"))
+@RepositoryRestResource
 public interface SubmissionRepository extends MongoRepository<Submission,String>{
 
     @Query(value="{ 'domain.name' : ?0 }")
+    @RestResource(rel="byDomain")
     Page<Submission> findByDomainName(@Param(value="domainName") String domainName, Pageable pageable);
 
-    @Query(value="{ '_id' : ?0, 'domain.name' : ?1 }")
-    Submission findOneByIdAndDomainName(@Param(value="id") String id, @Param(value="domainName") String domainName);
 
 }
